backport Revert "[Triggers] disable tests flaky on windows (#15955)"
